modified after AID
